src/pgtobq_sync/requirements.txt
psycopg2-binary
google-cloud-bigquery